- split up config.py into multiple files and config package - CheatConfig no longer relies on Config - first step of later Config overhaul
